Qualtrics round 1
This is what I've been trying to plug into qualtrics. I tried putting it into the header, but it was just dumping raw HTML all over the page. Putting the JS in a specific question fixed the raw html dump, but I'm getting a massive console logged error when I load the survey.